refactoring - pipelinerunstate
Creating a new file with all attributes of pipelinerunstate along with the
test file.

Refactored few functions as an attributes of pipelunerunstate
